DEVEX-283: Use the partial argument when calling removeFolder API methods
Summary:
The new "partial" argument passed to the "class-xxxx/removeFolder"
API call makes it possible to delete a folder containing more objects
than the system permits to delete at one time. This change adds "partial"
set to true to these API calls and reissues the calls until the "completed"
field returned in the response is set to "true", which indicates that
the folder with all the objects were deleted successfully.

Test Plan: run current integration tests

Reviewers: sking

Subscribers: sking

Differential Revision: https://dnanexus.phacility.com/D1522